Fix enhanced blind position update
