+++ src/accessories/blind.ts
@@ -1,31 +1,21 @@
 import { ComelitAccessory } from './comelit';
-import { BlindDeviceData, ComelitClient, OBJECT_SUBTYPE, ObjectStatus } from 'comelit-client';
+import { BlindDeviceData, ComelitClient, OBJECT_SUBTYPE } from 'comelit-client';
 import { ComelitPlatform } from '../comelit-platform';
-import { Callback, CharacteristicEventTypes, PlatformAccessory, Service } from 'homebridge';
+import { CharacteristicEventTypes, PlatformAccessory, Service } from 'homebridge';
 import { PositionState } from './hap';
-import Timeout = NodeJS.Timeout;
+import { CharacteristicSetCallback } from 'hap-nodejs';
 
-export class Blind extends ComelitAccessory<BlindDeviceData> {
+export abstract class Blind extends ComelitAccessory<BlindDeviceData> {
   static readonly OPEN = 100;
   static readonly CLOSED = 0;
 
   static readonly OPENING_CLOSING_TIME = 35; // 35 seconds to open approx. We should have this in the config
 
-  private coveringService: Service;
-  private timeout: Timeout;
-  private lastCommandTime: number;
-  private readonly closingTime: number;
-  private positionState: number;
+  protected coveringService: Service;
+  protected positionState: number;
 
-  constructor(
-    platform: ComelitPlatform,
-    accessory: PlatformAccessory,
-    client: ComelitClient,
-    closingTime?: number
-  ) {
+  constructor(platform: ComelitPlatform, accessory: PlatformAccessory, client: ComelitClient) {
     super(platform, accessory, client);
-    this.closingTime = (closingTime || Blind.OPENING_CLOSING_TIME) * 1000;
-    this.log.info(`Blind ${accessory.context.id} has closing time of ${this.closingTime}`);
   }
 
   protected initServices(): Service[] {
@@ -43,126 +33,17 @@ export class Blind extends ComelitAccessory<BlindDeviceData> {
 
     this.coveringService
       .getCharacteristic(Characteristic.TargetPosition)
-      .on(CharacteristicEventTypes.SET, async (position: number, callback: Callback) => {
-        if (this.device.sub_type === OBJECT_SUBTYPE.ENHANCED_ELECTRIC_BLIND) {
-          await this.enhancedBlindSet(position, callback);
-        } else {
-          await this.standardBlindSet(position, callback);
+      .on(
+        CharacteristicEventTypes.SET,
+        async (position: number, callback: CharacteristicSetCallback) => {
+          await this.setPosition(position, callback);
         }
-      });
+      );
 
     return [accessoryInformation, this.coveringService];
   }
 
-  private async enhancedBlindSet(position: number, callback: Callback) {
-    const Characteristic = this.platform.Characteristic;
-    try {
-      const currentPosition = this.coveringService.getCharacteristic(Characteristic.CurrentPosition)
-        .value as number;
-      this.log.info(`Setting position to ${position}%. Current position is ${currentPosition}`);
-      await this.client.setBlindPosition(this.device.id, Math.round(position * 2.55));
-      callback();
-    } catch (e) {
-      this.log.error(e.message);
-      callback(e);
-    }
-  }
+  public abstract setPosition(position: number, callback: CharacteristicSetCallback): Promise<void>;
 
-  private async standardBlindSet(position: number, callback: Callback) {
-    const Characteristic = this.platform.Characteristic;
-    try {
-      if (this.timeout) {
-        await this.resetTimeout();
-        callback();
-        return;
-      }
-
-      const currentPosition = this.coveringService.getCharacteristic(Characteristic.CurrentPosition)
-        .value as number;
-      const status = position < currentPosition ? ObjectStatus.OFF : ObjectStatus.ON;
-      const delta = currentPosition - position;
-      this.log.info(
-        `Setting position to ${position}%. Current position is ${currentPosition}. Delta is ${delta}`
-      );
-      if (delta !== 0) {
-        await this.client.toggleDeviceStatus(this.device.id, status);
-        this.lastCommandTime = new Date().getTime();
-        this.timeout = setTimeout(async () => {
-          return this.resetTimeout();
-        }, (this.closingTime * Math.abs(delta)) / 100);
-      }
-      callback();
-    } catch (e) {
-      this.log.error(e.message);
-      callback(e);
-    }
-  }
-  private async resetTimeout() {
-    // A timeout was set, this means that we are already opening or closing the blind
-    // Stop the blind and calculate a rough position
-    this.log.info(`Stopping blind`);
-    clearTimeout(this.timeout);
-    this.timeout = null;
-    await this.client.toggleDeviceStatus(
-      this.device.id,
-      this.positionState === PositionState.DECREASING ? ObjectStatus.ON : ObjectStatus.OFF
-    ); // stop the blind
-  }
-
-  public update(data: BlindDeviceData) {
-    const Characteristic = this.platform.Characteristic;
-    const status = parseInt(data.status);
-    const now = new Date().getTime();
-    switch (status) {
-      case ObjectStatus.ON:
-        this.lastCommandTime = now;
-        this.positionState = PositionState.INCREASING;
-        break;
-      case ObjectStatus.OFF: {
-        const position = this.positionFromTime();
-        this.lastCommandTime = 0;
-        this.log.info(
-          `Blind is now at position ${position} (it was ${
-            this.positionState === PositionState.DECREASING ? 'going down' : 'going up'
-          })`
-        );
-        this.positionState = PositionState.STOPPED;
-        this.coveringService.getCharacteristic(Characteristic.TargetPosition).updateValue(position);
-        this.coveringService
-          .getCharacteristic(Characteristic.CurrentPosition)
-          .updateValue(position);
-        this.coveringService
-          .getCharacteristic(Characteristic.PositionState)
-          .updateValue(PositionState.STOPPED);
-        break;
-      }
-      case ObjectStatus.IDLE:
-        this.lastCommandTime = now;
-        this.positionState = PositionState.DECREASING;
-        break;
-    }
-    this.log.info(
-      `Blind update: status ${status}, state ${this.positionState}, ts ${this.lastCommandTime}`
-    );
-  }
-
-  private positionFromTime() {
-    const Characteristic = this.platform.Characteristic;
-    const now = new Date().getTime();
-    // Calculate the number of milliseconds the blind moved
-    const delta = now - this.lastCommandTime;
-    const currentPosition = this.coveringService.getCharacteristic(Characteristic.CurrentPosition)
-      .value as number;
-    // Calculate the percentage of movement
-    const deltaPercentage = Math.round(delta / (this.closingTime / 100));
-    this.log.info(
-      `Current position ${currentPosition}, delta is ${delta} (${deltaPercentage}%). State ${this.positionState}`
-    );
-    if (this.positionState === PositionState.DECREASING) {
-      // Blind is decreasing, subtract the delta
-      return currentPosition - deltaPercentage;
-    }
-    // Blind is increasing, add the delta
-    return currentPosition + deltaPercentage;
-  }
+  public abstract update(data: BlindDeviceData);
 }
